requirements.txt
git+https://github.com/ultrafunkamsterdam/googletranslate>=2.2
google-cloud-translate>=3.6.1
langid>=1.1.6
lxml>=4.6.3
packaging>=21.3
Pillow>=9.0.0
psutil>=5.8.0
PySimpleGUI>=4.56.0
python-dateutil>=2.8.1
pytz>=2021.3
pywin32>=302
retry>=0.9.2
requests>=2.26.0
selenium>=4.1.0
timeago>=1.0.15
tzlocal>=4.1
undetected-chromedriver>=3.1.3